Initial token redemption options change
