+++ components/display-products.tsx
@@ -140,10 +140,14 @@ const DisplayEvents = ({
     if (focusedPubkey && productData.pubkey !== focusedPubkey) return;
     if (!productSatisfiesAllFilters(productData)) return;
 
-    if (productData.pubkey === "95a5e73109d4c419456372ce99bbf5823dfb6f77aed58d03f77ea052f150ee4a" && userPubkey !== productData.pubkey) {
+    if (
+      productData.pubkey ===
+        "95a5e73109d4c419456372ce99bbf5823dfb6f77aed58d03f77ea052f150ee4a" &&
+      userPubkey !== productData.pubkey
+    ) {
       return; // temp fix, add adult categories or separate from global later
     }
-    
+
     return (
       <ProductCard
         key={productData.id + "-" + index}
